ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.2",
    "torch>=2.9.1",
]
